+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 
 ## [Unreleased]
+### Added
+- Support for streaming ads with the Bitmovin Advertising Module (BAM) via `AdSourceType.BITMOVIN` in `AdvertisingConfig` ad schedules. Note: on iOS, VMAP is currently not supported with BAM.
 
 ## [1.10.0] - 2026-02-27